cephfs-shell: exit with non-zero on getting unrecognized cmd
Right now cephfs-shell on receiving unrecognized command prints an
appropriate message on stderr but the return value is zero. This is a
serious problem for users as well as for tests. It must exit with
non-zero value.

The return value chosen for this casei is 127 since bash does the
samething.