mds/MDSAuthCaps: allow adding multiple caps via "fs auth" cmd
When multiple caps are passed to "ceph fs authorize" command, the
command hangs because the MON crashes due to this command. Underlying
code assumes that only cap can be passed at a time by the user through
"ceph fs authorize" command. "ceph_assert()" checks if only caps is
passed, if not the code crashes.

Fixes: https://tracker.ceph.com/issues/64127
Signed-off-by: Rishabh Dave <ridave@redhat.com>