DNM: qa/cephfs: fix 2nd client in test
Test name: test_single_path_authorize_on_nonalphanumeric_fsname

Instead of leaving 2nd client mounted on a deleted CephFS, umount it
beforehand, remount it on new FS and check contents of "testfile"
through it too.

Interestingly this fixes bug and it can't be reproduced despite of
running it back to back for 5 times.

Signed-off-by: Rishabh Dave <ridave@redhat.com>