DNM: mgr/vol: allow "_nogroup" to opened by "ceph fs clone status" cmd
Command "ceph fs clone status" will need to open the internal default
group "_nogroup" to get the path of the snapshot which is being cloned
in order to create a  clone progress report. Therefore, use
CLONE_INTERNAL in helper methods of clone_status() and allow "_nogroup"
to be opened (only) in this case.

Signed-off-by: Rishabh Dave <ridave@redhat.com>