mgr/vol: create RTimer instance in stats_util only once
During developement, it seems code for RTimer instance creation was left
behind in method initiate_reporting() of class CloneProgressReporter due
to a mixup of patches during one of the rebases. Delete this line of code
in favour of same line of code which is present in __init__().

Following is the explanation on why this line of code is being deleted
in this commit from initiate_reporting() but not from __init__() -

Thread launched by RTimer instance is used to check and report progress
made by clone jobs. RTimer instance was initially being created in the
method inititate_reporting() of class CloneProgressReporter but later it
was moved to __init__() of same class. The rationale for this was to
create it in advance so that it is possible call RTimer.is_alive() to
check whether or not reporting has been initiated already. This is
necessary since calling RTimer.start() when RTimer thread is already
running will caused a crash.

initiate_reporting() is called everytime a clone is launched, this if
RTimer instance is created in this method, a new instance will be
created and assigned to a variable every time. Worse, old instance will
not have a variable referring to it and therefore will be automatically
deleted by Python. This probably has no effect on functionality from
user's POV but it is ineffecient due to unnecessary repetitive
creation/deletion every now and then.

Signed-off-by: Rishabh Dave <ridave@redhat.com>