DNM: mgr/vol: keep .meta for subvolume with retained snaps
Current metadata file for a subvolume is named ".meta". It is a symlink
to meta file for subvolume's current incarnation. By default in this
patch series this .meta file was unlinked to indicate that subvolume has
not current incarnation, it only has previous incarnations. However,
absence of ".meta" file leads SubvolumeBase.discover() to think that
this is legacy subvolume and that leads to creation of Subvolume v1
objects and running code v1. Ths is a severe bug.

So this patch ensure ".meta" file not unlinked until new incarnation is
created. This is not serious issue because ".meta" mentions inside that
current state in retained.

Signed-off-by: Rishabh Dave <ridave@redhat.com>